init: change js to ts
