+++ package.json
@@ -6,7 +6,6 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "cross-env NODE_ENV=production webpack",
-    "build:d": "cross-env NODE_ENV=development webpack",
     "start": "cross-env NODE_ENV=development webpack-dev-server --open",
     "format": "prettier --write .",
     "lint:js": "eslint ./src/**/*.js",
