+++ src/js/generate-game-screen-baby.ts
@@ -1,13 +1,11 @@
-import { showData } from './show-data.js';
-import cardsData from './../data/cards.json';
-
-import { generateStartScreen } from './generate-start-screen.js';
+import { showDataBaby } from './show-data-baby';
+import { generateStartScreen } from './generate-start-screen';
 
 /**
- * Генерация экрана игры
+ * Генерация экрана игры для малышей
  */
 
-export function generateGameScreen(container) {
+export function generateGameScreenBaby(container) {
   window.app.userTimeSekMin = [0, 0];
   window.app.guessedPairs = [];
   window.app.userCards = [];
@@ -59,7 +57,7 @@ export function generateGameScreen(container) {
   buttonRepeat.textContent = 'Начать заново';
 
   const cardsContainer = document.createElement('div');
-  cardsContainer.classList.add('cards-container');
+  cardsContainer.classList.add('cards-container-baby');
 
   container.appendChild(header);
   header.appendChild(timer);
@@ -69,16 +67,12 @@ export function generateGameScreen(container) {
   timer.appendChild(dot);
   timer.appendChild(sek);
 
-  header.appendChild(attempts);
-  attempts.appendChild(attempt1);
-  attempts.appendChild(attempt2);
-  attempts.appendChild(attempt3);
-
   header.appendChild(buttonRepeat);
   container.appendChild(cardsContainer);
 
-  generateCards(cardsContainer); // Вывод карт на экран
-  showData(cardsData);
+  generateCardsBaby(cardsContainer); // Вывод картинок на экран
+
+  showDataBaby();
 
   buttonRepeat.addEventListener('click', () => {
     container.innerHTML = '';
@@ -87,25 +81,36 @@ export function generateGameScreen(container) {
 }
 
 /**
- * Генерация и вывод карт на экран
+ * Генерация и вывод картинок на экран
  */
 
-function generateCards(container) {
-  const cardWidth = 95;
-  const cardHeight = 133;
+function generateCardsBaby(container) {
+  const cardsContainer = document.querySelector('.cards-container-baby');
+  cardsContainer.classList.add(`cards-container-baby_${window.app.level}`);
+
+  const numberOfCardsInRow = Number(window.app.level) + 1;
+  const numberOfCardsInColumn = 2;
+  const numberOfCards = numberOfCardsInRow * numberOfCardsInColumn;
+
+  const cardWidth = 150;
   const gap = 15;
-  const numberOfCardsInRow = 6;
+
+  if (window.app.level === '1') {
+    const timer = document.querySelector('.timer') as HTMLDivElement;
+    timer.style.marginRight = '70px';
+  }
+
+  for (let i = 0; i < numberOfCards; i++) {
+    const card = document.createElement('div');
+    card.classList.add('card', 'card_baby');
+    card.style.background = '#dddddd';
+    container.appendChild(card);
+  }
 
   container.style.width =
     numberOfCardsInRow * cardWidth + (numberOfCardsInRow - 1) * gap + 'px';
   container.style.height =
-    window.app.level * cardHeight + (window.app.level - 1) * gap + 'px';
-
-  for (let i = 0; i < numberOfCardsInRow * window.app.level; i++) {
-    const card = document.createElement('div');
-
-    card.classList.add('card');
-    card.style.background = '#c2f5ff';
-    container.appendChild(card);
-  }
+    numberOfCardsInColumn * cardWidth +
+    (numberOfCardsInColumn - 1) * gap +
+    'px';
 }
