+++ src/js/show-card-child.ts
@@ -1,27 +1,28 @@
-import { hideCardsBaby } from './hide-card.js';
-import { generateWinScreen } from './generate-win-screen.js';
+import { hideCardsChild } from './hide-card';
+import { generateWinScreen } from './generate-win-screen';
 
 /**
  * Открываем карту по клику
  */
 
-export function showCardBaby() {
-  const imgs = document.querySelectorAll('.card');
+export function showCardChild() {
+  const imgs = document.querySelectorAll('.card') as NodeListOf<HTMLDivElement>;
 
   for (let i = 0; i < imgs.length; i++) {
     imgs[i].addEventListener('click', () => {
+      imgs[
+        i
+      ].innerHTML = `<img class="img-child" src="${window.app.cardsForCurrentGame[i].img}" alt="${window.app.cardsForCurrentGame[i].name}" />`;
       imgs[i].style.transform = 'rotateY(180deg) scale(-1, 1)';
-      imgs[i].style.background = `${window.app.cardsForCurrentGame[i]}`;
-      imgs[i].style.border = '10px solid #ffffff';
 
       // добавляем пару выбранных карт в window.app.userCards
       if (!window.app.userCards[0] && !window.app.userTargets.includes(i)) {
-        window.app.userCards[0] = window.app.cardsForCurrentGame[i];
+        window.app.userCards[0] = window.app.cardsForCurrentGame[i].name;
         window.app.userTargets.push(i);
       }
 
       if (window.app.userCards[0] && !window.app.userTargets.includes(i)) {
-        window.app.userCards[1] = window.app.cardsForCurrentGame[i];
+        window.app.userCards[1] = window.app.cardsForCurrentGame[i].name;
         window.app.userTargets.push(i);
 
         //если карты не совпадают, обнуляем все результаты
@@ -30,7 +31,7 @@ export function showCardBaby() {
           window.app.userTargets = [];
           window.app.guessedPairs = [];
 
-          setTimeout(hideCardsBaby, window.app.commonDelay);
+          setTimeout(hideCardsChild, window.app.commonDelay);
         }
 
         //если карты совпадают, добавляем в список угаданных пар
@@ -44,7 +45,7 @@ export function showCardBaby() {
 
           window.app.userCards = [];
 
-          if (window.app.guessedPairs.length === Number(window.app.level) + 1) {
+          if (window.app.guessedPairs.length === window.app.level * 3) {
             clearInterval(window.app.stopwatch);
 
             setTimeout(generateWinScreen, window.app.commonDelay);
