+++ src/js/utils/arr-utils.ts
@@ -2,7 +2,7 @@
  *Дублирование элементов массива
  */
 
-export function doubleArr(a) {
+export function doubleArr(a: any[]) {
   return a.concat(a);
 }
 
@@ -10,7 +10,7 @@ export function doubleArr(a) {
  * Перемешивание элементов массива
  */
 
-export function shuffle(array) {
+export function shuffle(array: any[]) {
   for (let i = array.length - 1; i > 0; i--) {
     let j = Math.floor(Math.random() * (i + 1)); // случайный индекс от 0 до i
